Update API spec with `integer` type fixes

+++ internal/sdk/databasebranchpasswords.go
@@ -153,11 +153,11 @@ func (s *DatabaseBranchPasswords) ListPasswords(ctx context.Context, request ope
 		if err != nil {
 			return nil, err
 		}
-		var p float64 = 1
+		var p int64 = 1
 		if request.Page != nil {
 			p = *request.Page
 		}
-		nP := float64(p + 1)
+		nP := int64(p + 1)
 		r, err := ajson.Eval(b, "$.data")
 		if err != nil {
 			return nil, err
