+++ internal/sdk/databases.go
@@ -17,7 +17,7 @@ import (
 	"time"
 )
 
-// Databases -             Resources for managing databases within an organization.
+// Databases -           Resources for managing databases within an organization.
 type Databases struct {
 	rootSDK          *Planetscale
 	sdkConfiguration config.SDKConfiguration
@@ -153,11 +153,11 @@ func (s *Databases) ListDatabases(ctx context.Context, request operations.ListDa
 		if err != nil {
 			return nil, err
 		}
-		var p float64 = 1
+		var p int64 = 1
 		if request.Page != nil {
 			p = *request.Page
 		}
-		nP := float64(p + 1)
+		nP := int64(p + 1)
 		r, err := ajson.Eval(b, "$.data")
 		if err != nil {
 			return nil, err
