+++ internal/sdk/models/operations/createdatabase.go
@@ -44,7 +44,7 @@ type CreateDatabaseRequestBody struct {
 	// The database cluster size name (e.g., 'PS_10', 'PS_80'). Use the 'List available cluster sizes' endpoint to get available options for your organization. /v1/organizations/:organization/cluster-size-skus
 	ClusterSize string `json:"cluster_size"`
 	// The number of replicas for the database. 0 for non-HA, 2+ for HA.
-	Replicas *float64 `json:"replicas,omitzero"`
+	Replicas *int64 `json:"replicas,omitzero"`
 	// The kind of database to create.
 	Kind *KindRequest `json:"kind,omitzero"`
 	// For PostgreSQL databases, the PostgreSQL major version to use for the database. Defaults to the latest available major version.
@@ -72,7 +72,7 @@ func (c *CreateDatabaseRequestBody) GetClusterSize() string {
 	return c.ClusterSize
 }
 
-func (c *CreateDatabaseRequestBody) GetReplicas() *float64 {
+func (c *CreateDatabaseRequestBody) GetReplicas() *int64 {
 	if c == nil {
 		return nil
 	}
