+++ internal/sdk/models/operations/listbranches.go
@@ -50,9 +50,9 @@ type ListBranchesRequest struct {
 	// Order branches by created_at time
 	Order *Order `queryParam:"style=form,explode=true,name=order"`
 	// If provided, specifies the page offset of returned results
-	Page *float64 `default:"1" queryParam:"style=form,explode=true,name=page"`
+	Page *int64 `default:"1" queryParam:"style=form,explode=true,name=page"`
 	// If provided, specifies the number of returned results
-	PerPage *float64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
+	PerPage *int64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
 }
 
 func (l ListBranchesRequest) MarshalJSON() ([]byte, error) {
@@ -108,14 +108,14 @@ func (l *ListBranchesRequest) GetOrder() *Order {
 	return l.Order
 }
 
-func (l *ListBranchesRequest) GetPage() *float64 {
+func (l *ListBranchesRequest) GetPage() *int64 {
 	if l == nil {
 		return nil
 	}
 	return l.Page
 }
 
-func (l *ListBranchesRequest) GetPerPage() *float64 {
+func (l *ListBranchesRequest) GetPerPage() *int64 {
 	if l == nil {
 		return nil
 	}
