+++ internal/sdk/models/operations/listorganizationmembers.go
@@ -15,9 +15,9 @@ type ListOrganizationMembersRequest struct {
 	// Search term to filter members by name or email
 	Q *string `queryParam:"style=form,explode=true,name=q"`
 	// If provided, specifies the page offset of returned results
-	Page *float64 `default:"1" queryParam:"style=form,explode=true,name=page"`
+	Page *int64 `default:"1" queryParam:"style=form,explode=true,name=page"`
 	// If provided, specifies the number of returned results
-	PerPage *float64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
+	PerPage *int64 `default:"25" queryParam:"style=form,explode=true,name=per_page"`
 }
 
 func (l ListOrganizationMembersRequest) MarshalJSON() ([]byte, error) {
@@ -45,14 +45,14 @@ func (l *ListOrganizationMembersRequest) GetQ() *string {
 	return l.Q
 }
 
-func (l *ListOrganizationMembersRequest) GetPage() *float64 {
+func (l *ListOrganizationMembersRequest) GetPage() *int64 {
 	if l == nil {
 		return nil
 	}
 	return l.Page
 }
 
-func (l *ListOrganizationMembersRequest) GetPerPage() *float64 {
+func (l *ListOrganizationMembersRequest) GetPerPage() *int64 {
 	if l == nil {
 		return nil
 	}
