+++ internal/sdk/models/operations/updatebouncerresizerequest.go
@@ -13,7 +13,7 @@ type UpdateBouncerResizeRequestRequestBody struct {
 	// The bouncer size SKU name (e.g., 'PGB_5', 'PGB_10', 'PGB_20', 'PGB_40', 'PGB_80', 'PGB_160'). Defaults to 'PGB_5'.
 	BouncerSize *string `json:"bouncer_size,omitzero"`
 	// The number of PgBouncers per availability zone. Defaults to 1.
-	ReplicasPerCell *float64 `json:"replicas_per_cell,omitzero"`
+	ReplicasPerCell *int64 `json:"replicas_per_cell,omitzero"`
 	// Bouncer configuration parameters nested by namespace (e.g., {"pgbouncer": {"default_pool_size": "100"}}). Use the 'List cluster parameters' endpoint to retrieve available parameters. Only parameters with namespace 'pgbouncer' can be updated.
 	Parameters map[string]any `json:"parameters,omitzero"`
 }
@@ -36,7 +36,7 @@ func (u *UpdateBouncerResizeRequestRequestBody) GetBouncerSize() *string {
 	return u.BouncerSize
 }
 
-func (u *UpdateBouncerResizeRequestRequestBody) GetReplicasPerCell() *float64 {
+func (u *UpdateBouncerResizeRequestRequestBody) GetReplicasPerCell() *int64 {
 	if u == nil {
 		return nil
 	}
